moved all actors and corresponding tests to subpackage actor inside auction.domain
