+++ src/main/scala/io/scalac/auction/domain/DefaultAuctionService.scala
@@ -4,6 +4,7 @@ import akka.actor.typed.scaladsl.AskPattern._
 import akka.actor.typed.{ActorRef, Scheduler}
 import akka.stream.Materializer
 import akka.util.Timeout
+import io.scalac.auction.domain.actor.AuctionActorManager
 import io.scalac.auction.domain.model._
 import io.scalac.util.{ConfigProvider, ExecutionContextProvider, Logging}
